refactor(core): extract editDistance to a zero-dep leaf module (browser-port groundwork)
Stage-2 groundwork for the in-browser "Grade a repo" demo. The map of the audit
engine's browser port found the #1 blocker: core/linters.ts runs a
node:fs/process/glob side effect at IMPORT time (augmentToolPath), and the typo
detectors (tool-contract, hook-events) only import it for the pure `editDistance`
Levenshtein helper — so importing them into a browser bundle drags in the whole
linter-execution machinery.

Extract editDistance into src/core/edit-distance.ts (zero deps) and repoint
tool-contract + hook-events at it; linters.ts re-exports it for backward compat
and still uses it in closestRuleNames. No behavior change (tool-contract +
hook-events + scan suites green; build/tsc/lint/fmt clean).

Also captures the full engine-port build plan (the scanFiles seam, the file-map
= Set-membership finding, the remaining ncd/path/fs handling, and the 3-PR
sequence) in research/report-view-and-browser-demo.md so PRs 2-3 execute from it.
